feat: update test configuration to ignore transfer tests and rename test files for clarity
