+++ .pre-commit-config.yaml
@@ -27,6 +27,7 @@ repos:
         always_run: true
         args:
           - -x
+          - --ignore=tests/transfers
 
   # - repo: https://github.com/pre-commit/mirrors-mypy
   #   rev: v1.10.0  # Use the latest stable version or pin to your preference
