+++ tests/transfers/__init__.py
@@ -13,22 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from db import ThingContactAssociation
-from db.engine import session_ctx
-from transfers.contact_transfer import ContactTransfer
-from transfers.well_transfer import WellTransferer
-
-
-def test_multiple_wells():
-    pointids = ["MG-022", "MG-030", "MG-043"]
-    wt = WellTransferer(pointids=pointids)
-    wt.transfer()
-
-    ct = ContactTransfer(pointids=pointids)
-    ct.transfer()
-
-    with session_ctx() as sess:
-        assert sess.query(ThingContactAssociation).count() == 6
-
 
 # ============= EOF =============================================
